requirements/base.txt
-i https://pypi.python.org/simple
diff-match-patch==20121119
django-adminactions==1.5
django-ajax-selects==1.7.1
django-ajax-upload-widget==0.5.5
django-authtools==1.6.0
django-bootstrap3==11.0.0
django-braces==1.13.0
django-crispy-forms==1.7.2
django-environ==0.4.5
django-import-export==1.0.1
django-liststyle==0.1
django-simple-history==2.3.0
django-smart-selects==1.5.4
django-tinymce==2.7.0
django==2.1.1
easy-thumbnails==2.5
et-xmlfile==1.0.1
html5lib==1.0.1
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.7
pillow==5.2.0
pypdf2==1.26.0
pytz==2018.5
pyyaml==3.13
reportlab==3.5.6; python_version != '3.1.*'
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
webencodings==0.5.1
xhtml2pdf==0.2.3
xlrd==1.1.0
xlwt==1.3.0
